Separating build and config scripts
