+++ spfe-ot/1.0/scripts/python/spfelib/__init__.py
@@ -1,0 +1,2 @@
+from . import config
+from . import build
