--- spfe-ot/1.0/scripts/python/spfepy/__init__.py
@@ -1,1 +1,0 @@
-from .config import SPFEConfig
